pci: support 64-bit BAR sizes and upper-half probing
Change vfu_reg_info_t.size from uint32_t to uint64_t so that BARs
larger than 4GB (e.g. 16GB VRAM) produce correct size-mask responses
during PCI BAR probing.

Handle writes to the upper half of a 64-bit BAR pair (BAR[N] when
BAR[N-1].locatable==2): return the high 32 bits of the size mask
when the guest writes 0xFFFFFFFF for BAR size discovery.

+++ lib/private.h
@@ -93,8 +93,8 @@ struct migration;
 typedef struct  {
     /* Region flags, see VFU_REGION_FLAG_READ and friends. */
     uint32_t            flags;
-    /* Size of the region. */
-    uint32_t            size;
+    /* Size of the region (uint64_t to support >4GB BARs). */
+    uint64_t            size;
     /* Callback that is called when the region is read or written. */
     vfu_region_access_cb_t  *cb;
     /* Sparse mmap areas if set. */
